ELEM refactoring
Separated core/load and core/browser_type out from core/elem; allows cleaner requires.
